Change project configuration files and move them to Kotlin instead of Java source folder
